+++ lib/src/db/models.part.dart
@@ -176,6 +176,10 @@ class _Members extends Table {
 
   BoolColumn get invited => boolean().nullable()();
 
+  BoolColumn get banned => boolean().nullable()();
+
+  BoolColumn get shadowBanned => boolean().nullable()();
+
   BoolColumn get isModerator => boolean().nullable()();
 
   DateTimeColumn get createdAt => dateTime()();