+++ lib/src/db/offline_storage.g.dart
@@ -2843,6 +2843,8 @@ class _Member extends DataClass implements Insertable<_Member> {
   final DateTime inviteAcceptedAt;
   final DateTime inviteRejectedAt;
   final bool invited;
+  final bool banned;
+  final bool shadowBanned;
   final bool isModerator;
   final DateTime createdAt;
   final DateTime updatedAt;
@@ -2853,6 +2855,8 @@ class _Member extends DataClass implements Insertable<_Member> {
       this.inviteAcceptedAt,
       this.inviteRejectedAt,
       this.invited,
+      this.banned,
+      this.shadowBanned,
       this.isModerator,
       @required this.createdAt,
       this.updatedAt});
@@ -2874,6 +2878,10 @@ class _Member extends DataClass implements Insertable<_Member> {
           data['${effectivePrefix}invite_rejected_at']),
       invited:
           boolType.mapFromDatabaseResponse(data['${effectivePrefix}invited']),
+      banned:
+          boolType.mapFromDatabaseResponse(data['${effectivePrefix}banned']),
+      shadowBanned: boolType
+          .mapFromDatabaseResponse(data['${effectivePrefix}shadow_banned']),
       isModerator: boolType
           .mapFromDatabaseResponse(data['${effectivePrefix}is_moderator']),
       createdAt: dateTimeType
@@ -2902,6 +2910,12 @@ class _Member extends DataClass implements Insertable<_Member> {
     }
     if (!nullToAbsent || invited != null) {
       map['invited'] = Variable<bool>(invited);
+    }
+    if (!nullToAbsent || banned != null) {
+      map['banned'] = Variable<bool>(banned);
+    }
+    if (!nullToAbsent || shadowBanned != null) {
+      map['shadow_banned'] = Variable<bool>(shadowBanned);
     }
     if (!nullToAbsent || isModerator != null) {
       map['is_moderator'] = Variable<bool>(isModerator);
@@ -2932,6 +2946,11 @@ class _Member extends DataClass implements Insertable<_Member> {
       invited: invited == null && nullToAbsent
           ? const Value.absent()
           : Value(invited),
+      banned:
+          banned == null && nullToAbsent ? const Value.absent() : Value(banned),
+      shadowBanned: shadowBanned == null && nullToAbsent
+          ? const Value.absent()
+          : Value(shadowBanned),
       isModerator: isModerator == null && nullToAbsent
           ? const Value.absent()
           : Value(isModerator),
@@ -2954,6 +2973,8 @@ class _Member extends DataClass implements Insertable<_Member> {
       inviteAcceptedAt: serializer.fromJson<DateTime>(json['inviteAcceptedAt']),
       inviteRejectedAt: serializer.fromJson<DateTime>(json['inviteRejectedAt']),
       invited: serializer.fromJson<bool>(json['invited']),
+      banned: serializer.fromJson<bool>(json['banned']),
+      shadowBanned: serializer.fromJson<bool>(json['shadowBanned']),
       isModerator: serializer.fromJson<bool>(json['isModerator']),
       createdAt: serializer.fromJson<DateTime>(json['createdAt']),
       updatedAt: serializer.fromJson<DateTime>(json['updatedAt']),
@@ -2969,6 +2990,8 @@ class _Member extends DataClass implements Insertable<_Member> {
       'inviteAcceptedAt': serializer.toJson<DateTime>(inviteAcceptedAt),
       'inviteRejectedAt': serializer.toJson<DateTime>(inviteRejectedAt),
       'invited': serializer.toJson<bool>(invited),
+      'banned': serializer.toJson<bool>(banned),
+      'shadowBanned': serializer.toJson<bool>(shadowBanned),
       'isModerator': serializer.toJson<bool>(isModerator),
       'createdAt': serializer.toJson<DateTime>(createdAt),
       'updatedAt': serializer.toJson<DateTime>(updatedAt),
@@ -2982,6 +3005,8 @@ class _Member extends DataClass implements Insertable<_Member> {
           DateTime inviteAcceptedAt,
           DateTime inviteRejectedAt,
           bool invited,
+          bool banned,
+          bool shadowBanned,
           bool isModerator,
           DateTime createdAt,
           DateTime updatedAt}) =>
@@ -2992,6 +3017,8 @@ class _Member extends DataClass implements Insertable<_Member> {
         inviteAcceptedAt: inviteAcceptedAt ?? this.inviteAcceptedAt,
         inviteRejectedAt: inviteRejectedAt ?? this.inviteRejectedAt,
         invited: invited ?? this.invited,
+        banned: banned ?? this.banned,
+        shadowBanned: shadowBanned ?? this.shadowBanned,
         isModerator: isModerator ?? this.isModerator,
         createdAt: createdAt ?? this.createdAt,
         updatedAt: updatedAt ?? this.updatedAt,
@@ -3005,6 +3032,8 @@ class _Member extends DataClass implements Insertable<_Member> {
           ..write('inviteAcceptedAt: $inviteAcceptedAt, ')
           ..write('inviteRejectedAt: $inviteRejectedAt, ')
           ..write('invited: $invited, ')
+          ..write('banned: $banned, ')
+          ..write('shadowBanned: $shadowBanned, ')
           ..write('isModerator: $isModerator, ')
           ..write('createdAt: $createdAt, ')
           ..write('updatedAt: $updatedAt')
@@ -3026,9 +3055,13 @@ class _Member extends DataClass implements Insertable<_Member> {
                       $mrjc(
                           invited.hashCode,
                           $mrjc(
-                              isModerator.hashCode,
-                              $mrjc(createdAt.hashCode,
-                                  updatedAt.hashCode)))))))));
+                              banned.hashCode,
+                              $mrjc(
+                                  shadowBanned.hashCode,
+                                  $mrjc(
+                                      isModerator.hashCode,
+                                      $mrjc(createdAt.hashCode,
+                                          updatedAt.hashCode)))))))))));
   @override
   bool operator ==(dynamic other) =>
       identical(this, other) ||
@@ -3039,6 +3072,8 @@ class _Member extends DataClass implements Insertable<_Member> {
           other.inviteAcceptedAt == this.inviteAcceptedAt &&
           other.inviteRejectedAt == this.inviteRejectedAt &&
           other.invited == this.invited &&
+          other.banned == this.banned &&
+          other.shadowBanned == this.shadowBanned &&
           other.isModerator == this.isModerator &&
           other.createdAt == this.createdAt &&
           other.updatedAt == this.updatedAt);
@@ -3051,6 +3086,8 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
   final Value<DateTime> inviteAcceptedAt;
   final Value<DateTime> inviteRejectedAt;
   final Value<bool> invited;
+  final Value<bool> banned;
+  final Value<bool> shadowBanned;
   final Value<bool> isModerator;
   final Value<DateTime> createdAt;
   final Value<DateTime> updatedAt;
@@ -3061,6 +3098,8 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
     this.inviteAcceptedAt = const Value.absent(),
     this.inviteRejectedAt = const Value.absent(),
     this.invited = const Value.absent(),
+    this.banned = const Value.absent(),
+    this.shadowBanned = const Value.absent(),
     this.isModerator = const Value.absent(),
     this.createdAt = const Value.absent(),
     this.updatedAt = const Value.absent(),
@@ -3072,6 +3111,8 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
     this.inviteAcceptedAt = const Value.absent(),
     this.inviteRejectedAt = const Value.absent(),
     this.invited = const Value.absent(),
+    this.banned = const Value.absent(),
+    this.shadowBanned = const Value.absent(),
     this.isModerator = const Value.absent(),
     @required DateTime createdAt,
     this.updatedAt = const Value.absent(),
@@ -3085,6 +3126,8 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
     Expression<DateTime> inviteAcceptedAt,
     Expression<DateTime> inviteRejectedAt,
     Expression<bool> invited,
+    Expression<bool> banned,
+    Expression<bool> shadowBanned,
     Expression<bool> isModerator,
     Expression<DateTime> createdAt,
     Expression<DateTime> updatedAt,
@@ -3096,6 +3139,8 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
       if (inviteAcceptedAt != null) 'invite_accepted_at': inviteAcceptedAt,
       if (inviteRejectedAt != null) 'invite_rejected_at': inviteRejectedAt,
       if (invited != null) 'invited': invited,
+      if (banned != null) 'banned': banned,
+      if (shadowBanned != null) 'shadow_banned': shadowBanned,
       if (isModerator != null) 'is_moderator': isModerator,
       if (createdAt != null) 'created_at': createdAt,
       if (updatedAt != null) 'updated_at': updatedAt,
@@ -3109,6 +3154,8 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
       Value<DateTime> inviteAcceptedAt,
       Value<DateTime> inviteRejectedAt,
       Value<bool> invited,
+      Value<bool> banned,
+      Value<bool> shadowBanned,
       Value<bool> isModerator,
       Value<DateTime> createdAt,
       Value<DateTime> updatedAt}) {
@@ -3119,6 +3166,8 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
       inviteAcceptedAt: inviteAcceptedAt ?? this.inviteAcceptedAt,
       inviteRejectedAt: inviteRejectedAt ?? this.inviteRejectedAt,
       invited: invited ?? this.invited,
+      banned: banned ?? this.banned,
+      shadowBanned: shadowBanned ?? this.shadowBanned,
       isModerator: isModerator ?? this.isModerator,
       createdAt: createdAt ?? this.createdAt,
       updatedAt: updatedAt ?? this.updatedAt,
@@ -3146,6 +3195,12 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
     if (invited.present) {
       map['invited'] = Variable<bool>(invited.value);
     }
+    if (banned.present) {
+      map['banned'] = Variable<bool>(banned.value);
+    }
+    if (shadowBanned.present) {
+      map['shadow_banned'] = Variable<bool>(shadowBanned.value);
+    }
     if (isModerator.present) {
       map['is_moderator'] = Variable<bool>(isModerator.value);
     }
@@ -3167,6 +3222,8 @@ class _MembersCompanion extends UpdateCompanion<_Member> {
           ..write('inviteAcceptedAt: $inviteAcceptedAt, ')
           ..write('inviteRejectedAt: $inviteRejectedAt, ')
           ..write('invited: $invited, ')
+          ..write('banned: $banned, ')
+          ..write('shadowBanned: $shadowBanned, ')
           ..write('isModerator: $isModerator, ')
           ..write('createdAt: $createdAt, ')
           ..write('updatedAt: $updatedAt')
@@ -3255,6 +3312,32 @@ class $_MembersTable extends _Members with TableInfo<$_MembersTable, _Member> {
     );
   }
 
+  final VerificationMeta _bannedMeta = const VerificationMeta('banned');
+  GeneratedBoolColumn _banned;
+  @override
+  GeneratedBoolColumn get banned => _banned ??= _constructBanned();
+  GeneratedBoolColumn _constructBanned() {
+    return GeneratedBoolColumn(
+      'banned',
+      $tableName,
+      true,
+    );
+  }
+
+  final VerificationMeta _shadowBannedMeta =
+      const VerificationMeta('shadowBanned');
+  GeneratedBoolColumn _shadowBanned;
+  @override
+  GeneratedBoolColumn get shadowBanned =>
+      _shadowBanned ??= _constructShadowBanned();
+  GeneratedBoolColumn _constructShadowBanned() {
+    return GeneratedBoolColumn(
+      'shadow_banned',
+      $tableName,
+      true,
+    );
+  }
+
   final VerificationMeta _isModeratorMeta =
       const VerificationMeta('isModerator');
   GeneratedBoolColumn _isModerator;
@@ -3301,6 +3384,8 @@ class $_MembersTable extends _Members with TableInfo<$_MembersTable, _Member> {
         inviteAcceptedAt,
         inviteRejectedAt,
         invited,
+        banned,
+        shadowBanned,
         isModerator,
         createdAt,
         updatedAt
@@ -3349,6 +3434,16 @@ class $_MembersTable extends _Members with TableInfo<$_MembersTable, _Member> {
     if (data.containsKey('invited')) {
       context.handle(_invitedMeta,
           invited.isAcceptableOrUnknown(data['invited'], _invitedMeta));
+    }
+    if (data.containsKey('banned')) {
+      context.handle(_bannedMeta,
+          banned.isAcceptableOrUnknown(data['banned'], _bannedMeta));
+    }
+    if (data.containsKey('shadow_banned')) {
+      context.handle(
+          _shadowBannedMeta,
+          shadowBanned.isAcceptableOrUnknown(
+              data['shadow_banned'], _shadowBannedMeta));
     }
     if (data.containsKey('is_moderator')) {
       context.handle(