+++ lib/src/models/member.dart
@@ -25,8 +25,14 @@ class Member {
   /// The id of the interested user
   final String userId;
 
-  /// True if the user is a moderator of the channe;
+  /// True if the user is a moderator of the channel
   final bool isModerator;
+
+  /// True if the member is banned from the channel
+  final bool banned;
+
+  /// True if the member is shadow banned from the channel
+  final bool shadowBanned;
 
   /// The date of creation
   final DateTime createdAt;
@@ -45,6 +51,8 @@ class Member {
     this.isModerator,
     this.createdAt,
     this.updatedAt,
+    this.banned,
+    this.shadowBanned,
   });
 
   /// Create a new instance from a json
@@ -66,12 +74,16 @@ class Member {
     bool isModerator,
     DateTime createdAt,
     DateTime updatedAt,
+    bool banned,
+    bool shadowBanned,
   }) =>
       Member(
         user: user ?? this.user,
         inviteAcceptedAt: inviteAcceptedAt ?? this.inviteAcceptedAt,
         inviteRejectedAt: inviteRejectedAt ?? this.inviteRejectedAt,
         invited: invited ?? this.invited,
+        banned: banned ?? this.banned,
+        shadowBanned: shadowBanned ?? this.shadowBanned,
         role: role ?? this.role,
         userId: userId ?? this.userId,
         isModerator: isModerator ?? this.isModerator,